Added base.py
This python script contains the code for the expense tracker app, which can take the ITEM and its PRICE from the user and adds it to a database created in the same folder. The items are given a Serial Number and the date at which they are submitted to the db is assigned automatically. There is a date wise graph to show when the expenses were the most